+++ packages/terra-list/src/ListSection.jsx
@@ -200,9 +200,6 @@ const ListSection = ({
     style: getStyleforDrag(ListItem, snapshot, provider),
   });
 
-  /* eslint-disable-next-line no-param-reassign */
-  delete customProps?.isTabFocusDisabled;
-
   const renderSectionListItemsDom = () => (
     /* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
     <>